+++ dash_extensions/_imports_.py
@@ -1,9 +1,11 @@
 from .Download import Download
 from .Keyboard import Keyboard
+from .Linker import Linker
 from .Lottie import Lottie
 
 __all__ = [
     "Download",
     "Keyboard",
+    "Linker",
     "Lottie"
 ]