+++ src/lib/index.js
@@ -2,9 +2,11 @@
 import Download from './components/Download.react';
 import Lottie from './components/Lottie.react';
 import Keyboard from './components/Keyboard.react';
+import Linker from './components/Linker.react';
 
 export {
     Download,
     Lottie,
-    Keyboard
+    Keyboard,
+    Linker
 };